+++ core/lib/model/PreshaModelInterface.ts
@@ -1,4 +1,8 @@
+import { PreshaField } from '../types/PreshaSchema';
+
 export interface PreshaModelInterface<T> {
+  name: string;
+  fields: Record<string, PreshaField>;
   create(data: Partial<T>): Promise<T>;
   findById(id: string): Promise<T | null>;
   findAll(): Promise<T[]>;
